Move meter function to metrics.ts
