+++ packages/phone-number-privacy/signer/src/common/database/wrappers/domainRequest.ts
@@ -1,8 +1,7 @@
 import { DB_TIMEOUT, Domain, domainHash, ErrorMessage } from '@celo/phone-number-privacy-common'
 import Logger from 'bunyan'
 import { Knex } from 'knex'
-import { Histograms } from '../../../common/metrics'
-import { meter } from '../../web3/contracts'
+import { Histograms, meter } from '../../metrics'
 import {
   DomainRequestRecord,
   DOMAIN_REQUESTS_COLUMNS,
